moved some stuff over from abutil.
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@1687 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/libs/orcamisc/rand.cpp
@@ -3,6 +3,7 @@
 #include <stdlib.h>
 #include <assert.h>
 #include <iostream>
+#include <sstream>
 
 using namespace std; // for math functions in QNX
 
@@ -10,14 +11,14 @@ using namespace std; // for math functions in QNX
 
 namespace orcamisc {
 
-float
-normalRand( float mean, float std ) 
+double
+normalRand( double mean, double std ) 
 {
-   float x1, x2, w;
+   double x1, x2, w;
    do
    {
-      x1 = 2.0f * (float)rand( )/RAND_MAX - 1.0f;
-      x2 = 2.0f * (float)rand( )/RAND_MAX - 1.0f;
+      x1 = 2.0f * (double)rand( )/RAND_MAX - 1.0f;
+      x2 = 2.0f * (double)rand( )/RAND_MAX - 1.0f;
       w = x1*x1 + x2*x2;
    }
    while (w >= 1.0f || w < 1E-30);
